Add files via upload
The objective of the project was to predict the stock prices of Lockheed Martin from Historical Data taken from Yahoo Finance. The dates selected were from 09/21/2011 to 09/21/2023. I used 6 Machine Learning Models for predicting the stock prices. I checked for overfitting and the models' performances by plotting the Training and Validation errors and the importances of individual features. Finally I plotted the differences between the models and found out that Random Forest Regression was the most accurate in predicting the stock prices for Lockheed Martin.